# Conflicts:
#	include/villas/api/session.h
#	lib/api.c
#	lib/super_node.c
#	lib/web.c

+++ lib/hooks/fix_ts.c
@@ -29,7 +29,7 @@
 #include "timing.h"
 #include "sample.h"
 
-int fix_ts_read(struct hook *h, struct sample *smps[], size_t *cnt)
+int fix_ts_read(struct hook *h, struct sample *smps[], unsigned *cnt)
 {
 	struct timespec now;
